+++ GuitarAssistantProject/app/src/main/java/jenchenua/guitarassistantproject/fragments/Tab2.java
@@ -12,7 +12,7 @@ import jenchenua.guitarassistantproject.draw.FingeringDrawing;
 
 public class Tab2 extends Fragment {
     private FingeringDrawing fingering = null;
-    private int[] switches = {
+    private byte[] switches = {
             0, 1, 1, 0, 1,
             0, 0, 1, 0, 2,
             0, 1, 0, 1, 1,
